+++ substrate/client/rpc-spec-v2/src/archive/mod.rs
@@ -22,54 +22,11 @@
 //!
 //! Methods are prefixed by `archive`.
 
-use serde::{Deserialize, Serialize};
-
-/// The result of an RPC method.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-#[serde(untagged)]
-pub enum MethodResult {
-	/// Method generated a result.
-	Ok(MethodResultOk),
-	/// Method encountered an error.
-	Err(MethodResultErr),
-}
-
-impl MethodResult {
-	/// Constructs a successful result.
-	pub fn ok(result: impl Into<String>) -> MethodResult {
-		MethodResult::Ok(MethodResultOk { success: true, value: result.into() })
-	}
-
-	/// Constructs an error result.
-	pub fn err(error: impl Into<String>) -> MethodResult {
-		MethodResult::Err(MethodResultErr { success: false, error: error.into() })
-	}
-}
-
-/// The successful result of an RPC method.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MethodResultOk {
-	/// Method was successful.
-	pub success: bool,
-	/// The result of the method.
-	pub value: String,
-}
-
-/// The error result of an RPC method.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MethodResultErr {
-	/// Method encountered an error.
-	pub success: bool,
-	/// The error of the method.
-	pub error: String,
-}
-
 #[cfg(test)]
 mod tests;
 
 mod archive_storage;
+mod types;
 
 pub mod api;
 pub mod archive;
@@ -77,31 +34,4 @@ pub mod error;
 
 pub use api::ArchiveApiServer;
 pub use archive::Archive;
-
-mod tests {
-	use super::*;
-
-	#[test]
-	fn method_result_ok() {
-		let ok = MethodResult::ok("hello");
-
-		let ser = serde_json::to_string(&ok).unwrap();
-		let exp = r#"{"success":true,"value":"hello"}"#;
-		assert_eq!(ser, exp);
-
-		let ok_dec: MethodResult = serde_json::from_str(exp).unwrap();
-		assert_eq!(ok_dec, ok);
-	}
-
-	#[test]
-	fn method_result_error() {
-		let ok = MethodResult::err("hello");
-
-		let ser = serde_json::to_string(&ok).unwrap();
-		let exp = r#"{"success":false,"error":"hello"}"#;
-		assert_eq!(ser, exp);
-
-		let ok_dec: MethodResult = serde_json::from_str(exp).unwrap();
-		assert_eq!(ok_dec, ok);
-	}
-}
+pub use types::{MethodResult, MethodResultOk, MethodResultErr};
